+++ src/main/java/org/cyclops/integrateddynamics/api/network/IEnergyNetwork.java
@@ -4,7 +4,7 @@ package org.cyclops.integrateddynamics.api.network;
  * A network capability that holds energy.
  * @author rubensworks
  */
-public interface IEnergyNetwork extends IPositionedAddonsNetworkIngredients<Integer, Boolean> {
+public interface IEnergyNetwork extends IPositionedAddonsNetworkIngredients<Long, Boolean> {
 
     /**
      * @return The current network consumption rate.
